% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_statements.R
\name{simfinR_get_available_statements}
\alias{simfinR_get_available_statements}
\title{Get available financial statements from simfin}
\usage{
simfinR_get_available_statements(id_sim, api_key, silent = TRUE)
}
\arguments{
\item{id_sim}{Id of company (see \code{\link{simfinR_get_available_companies}} for downloading a table with ids)}

\item{api_key}{Your api key}

\item{silent}{Be silent? (TRUE or FALSE)}
}
\value{
A dataframe with available statements
}
\description{
Uses the api to get available statements for a single company.
}
\examples{
\dontrun{
id_companies <- 59265
api_key <- 'YOURAPIHERE'

df_available <- simfinR_get_available_statements(id_companies, api_key = api_key)
}
}
