% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_company.R
\name{simfinR_get_info_company}
\alias{simfinR_get_info_company}
\title{Get description information for a company}
\usage{
simfinR_get_info_company(id_sim, api_key, cache_folder)
}
\arguments{
\item{id_sim}{Id of company}

\item{api_key}{Your api key}

\item{cache_folder}{Folder to save cache files}
}
\value{
A list with information
}
\description{
Uses the simfin api to get information for a single company.
}
\examples{
\dontrun{
api_key <- 'YOURAPIHERE'
l_info <- simfinR_get_info_company(59265, api_key, tempdir())
}
}
