% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{simfinR_get_price_data}
\alias{simfinR_get_price_data}
\title{Get price data for a group of companies}
\usage{
simfinR_get_price_data(id_companies, api_key, cache_folder = "simfin_cache")
}
\arguments{
\item{id_companies}{A vector of ids of companies}

\item{api_key}{Your api key}

\item{cache_folder}{The cache folder to save files}
}
\value{
A dataframe with price data
}
\description{
This function will use the simfin api to grab adjusted prices of stocks.
}
\examples{
\dontrun{
df_prices <- simfinR_get_price_data(59265, 'YOURAPIKEY')
}
}
