% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_share_classes.R
\name{simfinR_get_share_classes}
\alias{simfinR_get_share_classes}
\title{Get share class data for a group of companies}
\usage{
simfinR_get_share_classes(id_companies, api_key, cache_folder = "simfin_cache")
}
\arguments{
\item{id_companies}{A vector of ids of companies}

\item{api_key}{Your api key}

\item{cache_folder}{The cache folder to save files}
}
\value{
A dataframe with share class data
}
\description{
This function will use the simfin api to grab share classes of stocks.
}
\examples{
\dontrun{
df <- simfinR_get_share_classes(59265, 'YOURAPIKEY')
}
}
