% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aggregated_shares_outstanding.R
\name{simfinR_get_single_aggregated_shares_outstanding}
\alias{simfinR_get_single_aggregated_shares_outstanding}
\title{Get aggregated shares outstanding data for single company}
\usage{
simfinR_get_single_aggregated_shares_outstanding(
  id_sim,
  api_key,
  filter = "all",
  cache_folder = "simfin_cache"
)
}
\arguments{
\item{id_sim}{The simfin id}

\item{api_key}{Your api key}

\item{filter}{Omit to return all types of shares outstanding, or one of "common-outstanding", "common-outstanding-basic", "common-outstanding-diluted", "preferred". See https://simfin.com/api/v1/documentation/#operation/getCompSharesAggregated for more detail.}

\item{cache_folder}{Cache folder to save files}
}
\value{
A dataframe with aggregated shares outstanding
}
\description{
Helper function of \code{\link{simfinR_get_aggregated_shares_outstanding}}.
Will fetch data for a single company and save results in cache.
}
\examples{
\dontrun{
df <- simfinR_get_single_aggregated_shares_outstanding(59265, 'YOURAPIKEY')
}
}
