% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{simfinR_get_single_price_data}
\alias{simfinR_get_single_price_data}
\title{Get price data for single company}
\usage{
simfinR_get_single_price_data(id_sim, api_key, cache_folder = "simfin_cache")
}
\arguments{
\item{id_sim}{The simfin id}

\item{api_key}{Your api key}

\item{cache_folder}{Cache folder to save files}
}
\value{
A dataframe with adjusted prices
}
\description{
Helper function of \code{\link{simfinR_get_price_data}}.
Will fetch data for a single company and save results in cache.
}
\examples{
\dontrun{
df_prices <- simfinR_get_single_price_data(59265, 'YOURAPIKEY')
}
}
