% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data_share_classes.R
\name{simfinR_get_single_price_for_class}
\alias{simfinR_get_single_price_for_class}
\title{Get price data by share class for single company and share class}
\usage{
simfinR_get_single_price_for_class(
  id_sim,
  id_share_class,
  api_key,
  cache_folder = "simfin_cache"
)
}
\arguments{
\item{id_sim}{The simfin id}

\item{id_share_class}{The simfin share class id from simfinR_get_share_classes}

\item{api_key}{Your api key}

\item{cache_folder}{Cache folder to save files}
}
\value{
A dataframe with price data for the share class
}
\description{
Helper function of \code{\link{simfinR_get_single_price_by_class}}.
Will fetch data for a single company and save results in cache.
}
\examples{
\dontrun{
df <- simfinR_get_single_shares_outstanding_for_class(59265, 'YOURAPIKEY')
}
}
