% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_share_classes.R
\name{simfinR_get_single_share_classes}
\alias{simfinR_get_single_share_classes}
\title{Get share class data for single company}
\usage{
simfinR_get_single_share_classes(
  id_sim,
  api_key,
  cache_folder = "simfin_cache"
)
}
\arguments{
\item{id_sim}{The simfin id}

\item{api_key}{Your api key}

\item{cache_folder}{Cache folder to save files}
}
\value{
A dataframe with share classes
}
\description{
Helper function of \code{\link{simfinR_get_share_classes}}.
Will fetch data for a single company and save results in cache.
}
\examples{
\dontrun{
df_share_classes <- get_share_classes(59265, 'YOURAPIKEY')
}
}
