% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simfinR_id_to_name}
\alias{simfinR_id_to_name}
\title{Finds a company name by its simfin id}
\usage{
simfinR_id_to_name(id_in, api_key)
}
\arguments{
\item{id_in}{Id of company}

\item{api_key}{Your api key}
}
\value{
A string (name of company)
}
\description{
Finds a company name by its simfin id
}
\examples{
\dontrun{
api_key <- 'YOURAPIHERE'
my_name <- simfinR_id_to_name(59265, api_key)
}
}
