% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simfinR_period_to_date}
\alias{simfinR_period_to_date}
\title{Period to date}
\usage{
simfinR_period_to_date(period_in, year_in)
}
\arguments{
\item{period_in}{The period ('Q1', 'Q2', 'Q3', 'Q4', 'FY')}

\item{year_in}{The year}
}
\value{
A single date object
}
\description{
Returns the last date for a given period and year.
}
\examples{
first_quarter <- simfinR_period_to_date('Q1', 2019)
}
