% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simfinR_set_fromJSON}
\alias{simfinR_set_fromJSON}
\title{Sets memoise cache function}
\usage{
simfinR_set_fromJSON(cache_folder = tempdir())
}
\arguments{
\item{cache_folder}{The local folder where to save files}
}
\value{
A memoised function
}
\description{
USes the memoise package to set a cache function for jsonlite::fromJSON
The choice of the cache folder defines where to save the local files. If not set, will use tempdir()
}
\examples{
mem_fromJSON <- simfinR_set_fromJSON()

}
